% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_consonance_test}
\alias{add_consonance_test}
\title{augment a consonance suite with a new consonance test}
\usage{
add_consonance_test(suite, test)
}
\arguments{
\item{suite}{object of class consonance_suite}

\item{test}{object of class consonance_test}
}
\value{
consonance suite
}
\description{
augment a consonance suite with a new consonance test
}
\examples{


}
