% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_test}
\alias{add_test}
\title{add a new test into a consonance test suite}
\usage{
add_test(suite, .desc, .fun, ...)
}
\arguments{
\item{suite}{object of class consonance_suite}

\item{.desc}{character}

\item{.fun}{function, must accept at least one argument}

\item{...}{other parameters, passed to constructor consonance_test}
}
\value{
consonance suite
}
\description{
add a new test into a consonance test suite
}
