% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{attach_consonance}
\alias{attach_consonance}
\title{attach a consonance object to an existing model}
\usage{
attach_consonance(object, suite)
}
\arguments{
\item{object}{list-like object, for example a lm or glm model}

\item{suite}{object of class consonance_suite}
}
\value{
list-like object with a consonance component
}
\description{
attach a consonance object to an existing model
}
\examples{

# create a regression model and a suite of tests
xy <- data.frame(x=1:10, y=1:10)
xy.lm <- lm(y~x, data=xy)
suite <- consonance_suite() +
   consonance_assert("x numeric", assert_numeric)
xy.model = attach_consonance(xy.lm, suite)

# preview that the test suite is attached to the model
xy.model$consonance

# use the model as-if it were a test suite
test_consonance(xy, xy.model)

}
