% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consonance_model_data}
\alias{consonance_model_data}
\title{synthetic dataset with five variables plus an outcome variable}
\format{
An object of class \code{data.frame} with 200 rows and 6 columns.
}
\usage{
data(consonance_model_data)

# @format data frame with six columns: \code{y}, \code{x1}, \code{x2},
}
\description{
The dataset is used in the vignette to illustrate consonance testing
in the context of multiple regression
}
\keyword{datasets}
