% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{consonance_suite}
\alias{consonance_suite}
\title{create a new suite of consonance tests}
\usage{
consonance_suite(
  level = c("error", "warning"),
  logging.level = c("WARN", "INFO", "ERROR"),
  log.file = NULL
)
}
\arguments{
\item{level}{character, determines when to halt execution}

\item{logging.level}{character, determines level of verbosity in logs.
This argument is over-ridden by setting a non-default logger.}

\item{log.file}{character, path for a log file, or leave NULL to log
to the console
The default "log4r" sets up a simple console logger using package log4r.
Setting a non-default logger over-rides values set in logging.level.}
}
\value{
object of class consonance
}
\description{
create a new suite of consonance tests
}
\examples{

suite <- consonance_suite()
suite

}
