% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach.R
\name{detach_consonance}
\alias{detach_consonance}
\title{detach a suite from an object}
\usage{
detach_consonance(object)
}
\arguments{
\item{object}{an object that includes a consonance suite}
}
\value{
consonance suite that is not attached to an object
}
\description{
detach a suite from an object
}
\examples{

# first attach a suite to a model
xy <- data.frame(x=1:10, y=1:10)
xy.lm <- lm(y~x, data=xy)
suite <- consonance_suite() +
   consonance_test("x numeric", is.numeric)
xy.model <- attach_consonance(xy.lm, suite)
xy.model$consonance

# detach the suite from the model
suite_2 <- detach_consonance(xy.model)

# note: to obtain the model without the suite, set the component to NULL
xy.model$consonance <- NULL

}
