% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{match.consonance.arg}
\alias{match.consonance.arg}
\title{variation on match.arg}
\usage{
match.consonance.arg(arg, label)
}
\arguments{
\item{arg}{character, value supplied by user}

\item{label}{character, identifier to a list that holds allowed choices}
}
\value{
value converted to the allowed values
}
\description{
this emits a more informative error message
it also checks against consonance_defaults
}
