% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{+.consonance}
\alias{+.consonance}
\title{concatenate consonance tests into a consonance suite}
\usage{
\method{+}{consonance}(e1, e2)
}
\arguments{
\item{e1}{consonance_suite or consonance_test}

\item{e2}{consonance_suite or consonance_test}
}
\value{
a suite of consonance tests
}
\description{
concatenate consonance tests into a consonance suite
}
\examples{

# declare two consonance tests for a hypothetical
# data frame with columns "x" and "y"
test_x <- consonance_test("x numeric", is.numeric, .var="x")
test_y <- consonance_test("y numeric", is.numeric, .var="y")

# create a suite by adding two tests to an empty suite
suite_2 <- consonance_suite() +  test_x + test_y

# create a suite by adding two tests together
# (tests automatically grouped into a suite)
suite_2 <- test_x + test_y

# create a large suite by concatenating two suite together
suite_4 <- suite_2 + suite_2
length(suite_4$tests)
# (note that + really means concatentation and tests can be redundant)

}
