% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{+.consonance_test}
\alias{+.consonance_test}
\title{concatenate consonance tests into a consonance suite}
\usage{
\method{+}{consonance_test}(e1, e2)
}
\arguments{
\item{e1}{object of class consonance or consonance_test}

\item{e2}{object of class consonance_test}
}
\value{
a suite of consonance tests
}
\description{
concatenate consonance tests into a consonance suite
}
