% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.consonance_suite}
\alias{print.consonance_suite}
\title{Display a summary for a consonance test suite}
\usage{
\method{print}{consonance_suite}(x, ...)
}
\arguments{
\item{x}{object of class consonance}

\item{...}{other arguments ignored}
}
\description{
Display a summary for a consonance test suite
}
\keyword{internal}
