% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.consonance_test}
\alias{print.consonance_test}
\title{Display a summary for a consonance test}
\usage{
\method{print}{consonance_test}(x, ...)
}
\arguments{
\item{x}{object of class consonance_test}

\item{...}{other arguments ignored}
}
\description{
Display a summary for a consonance test
}
\keyword{internal}
