% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_consonance}
\alias{test_consonance}
\title{check consonance of an object with a suite of tests}
\usage{
test_consonance(x, suite, skip = FALSE, logging.level = NULL, log.file = NA)
}
\arguments{
\item{x}{data object}

\item{suite}{object of class consonance}

\item{skip}{logical, set to TRUE to skip all tests}

\item{logging.level}{character, code to indicate logging level,
leave NULL to use the logging level specified in suite constructor}

\item{log.file}{character, path to log file, or leave NA to follow
the logger set up within suite}
}
\value{
object x, unchanged
}
\description{
The function always returns its primary input and should be invoked
for its side effects. It stops when a data
}
\examples{



}
