% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\alias{validate_inner}
\title{check consonance of an object with a suite of tests}
\usage{
validate(
  x,
  suite,
  level = NA,
  skip = FALSE,
  skip.action = c("log", "none"),
  logging.level = c(NA, "INFO", "WARN", "ERROR", "NONE"),
  log.file = NA,
  x.name = NA,
  suite.name = NA
)

validate_inner(
  x,
  suite,
  level = NA,
  skip = FALSE,
  logging.level = NA,
  log.file = NA,
  x.name = NA,
  suite.name = NA,
  ...
)
}
\arguments{
\item{x}{data object}

\item{suite}{object of class consonance}

\item{level}{character, strictness level, use "warning" or "error" to
determine the minimal criteria to halt execution}

\item{skip}{logical, set to TRUE to skip all tests}

\item{skip.action}{character, determines what happens when skip=TRUE.
Action 'log' sends an INFO message to the logger, 'none' is silent.}

\item{logging.level}{character, code to indicate logging level;
leave NA to use the logging level specified in suite constructor;
set NULL to avoid logging altogether}

\item{log.file}{character, path to log file, or leave NA to follow
the logger set up within suite}

\item{x.name}{character, (used by validate_batch), leave NA during
standard usage}

\item{suite.name}{character, (used by validate_batch), leave NA during
standard usage}

\item{...}{other arguments, not used}
}
\value{
object x, unchanged

vector with pass, warning, error, and stop counts
}
\description{
The function always returns its primary input and should be invoked
for its side effects. It stops when a data
}
\examples{

# create a test or a suite with a test
test_1 <- consonance_test("a numeric", is.numeric)
suite_1 <- consonance_suite() + test_1

# applying the test or suite on numeric data
# should execute without visible output
validate(c(1, 2, 3), test_1)
validate(c(1, 2, 3), suite_1)

# applying on non-numeric data should produce messages and stop execution
# validate(letters, test_1)
# validate(letters, suite_1)

}
\keyword{internal}
