% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_batch}
\alias{validate_batch}
\title{applies validate() to a series of objects and consonance suites}
\usage{
validate_batch(objects, suites, envir, ...)
}
\arguments{
\item{objects, }{character vector with data object names}

\item{envir, }{environment from which to fetch data and suites}

\item{...}{other arguments passed to validate()}

\item{suies, }{character vector with suite names}
}
\value{
vectoring with pass, warning, error, stop counts.
The most reliable value is "stop" that signals whether,
overall, the batch generated a stop condition
}
\description{
applies validate() to a series of objects and consonance suites
}
\keyword{internal}
